requirements.txt
streamlit>=1.35.0
yfinance>=0.2.40
pandas>=2.0.0
plotly>=5.20.0
pyportfolioopt>=1.5.5
groq>=0.9.0
python-dotenv>=1.0.0
numpy
feedparser>=6.0.0